requirements.txt
absl-py==0.2.2
adversarial-vision-challenge==1.0.5
alabaster==0.7.10
anaconda-client==1.6.9
anaconda-navigator==1.7.0
anaconda-project==0.8.2
asgiref==3.2.7
asn1crypto==0.24.0
astor==0.8.1
astroid==1.6.1
astropy==2.0.3
attrs==17.4.0
Babel==2.5.3
backports.functools-lru-cache==1.5
backports.shutil-get-terminal-size==1.0.0
bandmat==0.7
beautifulsoup4==4.6.0
bitarray==0.8.1
bkcharts==0.2
blaze==0.11.3
bleach==2.1.2
blinker==1.4
bokeh==0.12.13
boto==2.48.0
boto3==1.7.45
botocore==1.10.84
Bottleneck==1.2.1
bson==0.5.9
bz2file==0.98
cachetools==4.0.0
certifi==2018.1.18
cffi==1.11.4
chardet==3.0.4
charset-normalizer==2.0.12
cleverhans==2.1.0
click==6.7
cloudpickle==1.3.0
clyent==1.2.2
colorama==0.3.9
configparser==5.0.2
contextlib2==0.5.5
crcmod==1.7
crowdai-api==0.1.22
crowdai-repo2docker==0.5.0
cryptography==2.1.4
cycler==0.10.0
cymem==1.31.2
Cython==0.28.3
cytoolz==0.9.0
dask==0.16.1
datashape==0.5.4
deap==1.2.2
decorator==4.2.1
dill==0.2.9
distributed==1.20.2
Django==3.0.5
dm-sonnet==1.23
docker==4.2.0
docutils==0.14
eagerpy==0.27.0
easydict==1.7
entrypoints==0.2.3
enum34==1.1.10
ephemeral-port-reserve==1.1.4
escapism==1.0.0
esdk-obs-python==3.21.4
et-xmlfile==1.0.1
fastcache==1.0.2
filelock==2.0.13
Flask==1.0.2
Flask-Cors==3.0.3
foolbox==3.0.0
future==0.18.2
futures==3.1.0
gast==0.3.3
gensim==3.4.0
gevent==1.2.2
gitdb==4.0.2
GitPython==3.1.0
glob2==0.6
gmpy2==2.0.8
google-api-core==1.16.0
google-api-python-client==1.8.0
google-auth==1.13.1
google-auth-httplib2==0.0.3
googleapis-common-protos==1.51.0
graphviz==0.8.3
greenlet==0.4.12
grpcio==1.28.1
gunicorn==19.8.1
gym==0.17.1
h5py==2.8.0
HeapDict==1.0.0
html5lib==1.0.1
httplib2==0.17.1
huaweicloudsdkcore==3.0.50
idna==2.6
image==1.5.24
imageio==2.3.0
imagesize==0.7.1
imgaug==0.2.6
importlib-metadata==4.11.3
intervaltree==2.1.0
ipykernel==4.8.0
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.1.1
isort==4.2.15
itsdangerous==0.24
jdcal==1.3
jedi==0.11.1
jieba==0.39
Jinja2==2.10
jmespath==0.9.5
jsonschema==2.6.0
jupyter-client==5.1.0
jupyter-core==4.3.0
jupyterlab==0.35.4
jupyterlab-server==0.2.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.2.0
lazy-object-proxy==1.3.1
llvmlite==0.21.0
lmdb==0.94
locket==0.2.0
lxml==4.1.1
Markdown==3.2.1
MarkupSafe==1.0
matplotlib==2.2.3
mccabe==0.6.1
mistune==0.8.3
mnist==0.2.2
mock==4.0.2
modelarts==1.4.2
moxing-framework==1.17.3
moxing-tensorflow==1.15.1
mpmath==1.0.0
msgpack==0.6.2
msgpack-numpy==0.4.4.3
msgpack-python==0.5.1
multipledispatch==0.4.9
murmurhash==0.28.0
navigator-updater==0.1.0
nbconvert==5.3.1
nbformat==4.4.0
netifaces==0.10.7
networkx==2.1
nibabel==2.3.1
nltk==3.2.5
nose==1.3.7
notebook==5.0.0
numba==0.36.2
numexpr==2.6.4
numpy==1.19.1
numpydoc==0.7.0
oauth2client==4.1.3
oauthlib==3.0.1
odo==0.5.1
olefile==0.45.1
onnx==1.0.1
openpyxl==2.4.10
openslide-python==1.1.1
packaging==16.8
pandas==0.22.0
pandocfilters==1.4.2
parso==0.1.1
partd==0.3.8
path.py==10.5
pathlib2==2.3.0
patsy==0.5.0
pep8==1.7.1
pexpect==4.3.1
pickleshare==0.7.4
Pillow==5.0.0
pkginfo==1.4.1
plac==0.9.6
pluggy==0.6.0
ply==3.10
preshed==1.0.1
prettytable==2.1.0
prompt-toolkit==1.0.15
protobuf==3.11.3
psutil==5.7.2
psycopg2==2.7.4
ptyprocess==0.5.2
py==1.5.2
pyarrow==0.16.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.3.1
pycosat==0.6.3
pycparser==2.18
pycrypto==2.6.1
pydicom==1.2.0
pyflakes==1.6.0
pyglet==1.5.0
Pygments==2.2.0
pygtrie==2.3
PyJWT==1.7.1
pylint==1.8.2
PyMySQL==0.8.1
pyodbc==4.0.22
pyOpenSSL==17.5.0
pyparsing==2.2.0
PySocks==1.6.7
pytest==3.3.2
pytest-runner==4.2
python-crfsuite==0.9.6
python-dateutil==2.6.1
python-docx==0.8.7
python-gitlab==2.2.0
python-json-logger==0.1.11
python-prctl==1.7
pytz==2017.3
PyWavelets==0.5.2
PyYAML==3.12
pyzmq==17.0.0
QtAwesome==0.4.4
qtconsole==4.3.1
QtPy==1.3.1
rawpy==0.11.0
redis==3.4.1
regex==2017.4.5
requests==2.18.4
requests-futures==1.0.0
requests-oauthlib==1.0.0
rope==0.10.7
rsa==4.0
ruamel.yaml==0.15.35
s3transfer==0.1.13
scikit-image==0.14.5
scikit-learn==0.19.1
scipy==1.2.2
seaborn==0.8.1
semantic-version==2.8.5
Send2Trash==1.4.2
setproctitle==1.1.10
simplegeneric==0.8.1
SimpleITK==1.1.0
six==1.11.0
smart-open==1.5.3
smmap==3.0.1
snowballstemmer==1.2.1
sortedcollections==0.5.3
sortedcontainers==1.5.9
spacy==2.0.12
Sphinx==1.6.6
sphinxcontrib-websupport==1.0.1
spyder==3.2.6
SQLAlchemy==1.2.1
sqlparse==0.3.1
statsmodels==0.8.0
sympy==1.1.1
tables==3.4.2
tabulate==0.8.7
tblib==1.3.2
tenacity==8.0.1
tensor2tensor==1.9.0
tensorboard==1.13.1
tensorboard-logger==0.1.0
tensorflow-estimator==1.13.0
tensorflow-gpu==1.13.1
tensornets==0.3.5
tensorpack==0.8.8
termcolor==1.1.0
terminado==0.8.1
testpath==0.3
tflearn==0.3.2
thinc==6.10.3
toolz==0.9.0
Top==1.0.0
tornado==4.5.3
tqdm==4.45.0
traitlets==4.3.2
twython==3.7.0
typing==3.6.2
ujson==1.35
unicodecsv==0.14.1
uritemplate==3.0.1
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==0.14.1
widgetsnbextension==3.1.0
wrapt==1.10.11
xlrd==1.1.0
XlsxWriter==1.0.2
xlwt==1.3.0
zict==0.1.3
zipp==3.7.0
